WIP - add turret manager and factory

+++ Assets/ScriptableObjects/Data/Turret Types/AttackTurretSO.cs
@@ -5,6 +5,6 @@ public class AttackTurretSO : BaseTurretSO
 {
     public float shootSpeed;
     public int attackDamage;
-    public int attackRange;
+    public float attackRange;
     public float attackDelay = 0.2f;
 }
